fixed conversion warnings

+++ dtnn_clean/prediction_blob.cpp
@@ -7,7 +7,7 @@ namespace dtnn {
   PredictionBatch& PredictionBlob::batch(dim_t size) {
     if (location_ + size >= inputs_.dims(3))
       location_ = 0;
-    current_batch_.inputs = inputs_(af::span, af::span, af::span, af::seq(location_, size - 1));
+    current_batch_.inputs = inputs_(af::span, af::span, af::span, af::seq((double)location_, (double)size - 1));
     return current_batch_;
   }
   af::dim4 PredictionBlob::input_dimensions() {
